# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"time"

	"github.com/google/uuid"
)

type File struct {
	Filename  uuid.UUID
	OwnerID   int32
	CreatedAt time.Time
	UpdatedAt time.Time
}

type User struct {
	ID           int32
	PasswordHash string
	Name         string
	CreatedAt    time.Time
	UpdatedAt    time.Time
}